[Feat] #23-Set CollectionView Header
